+++ src/formatter/rufo.ts
@@ -11,6 +11,10 @@ const DEFAULT_OPTIONS: RufoOptions = {
   useBundler: false
 };
 
+function cleanUpError(message: string) {
+  return message.replace('STDIN is invalid code. ', '')
+}
+
 export default class Rufo {
   public test(): Promise<void> {
     return new Promise((resolve, reject) => {
@@ -60,8 +64,8 @@ export default class Rufo {
 
       const rufo = this.spawn(args);
 
-      if (rufo.stdin === null || rufo.stdout === null) {
-        const msg = "Couldn't initialize STDIN or STDOUT";
+      if (rufo.stdin === null || rufo.stdout === null || rufo.stderr === null) {
+        const msg = "Couldn't initialize STDIN, STDOUT or STDERR";
         console.warn(msg);
         vscode.window.showErrorMessage(msg);
         reject(msg);
@@ -69,6 +73,7 @@ export default class Rufo {
       }
 
       let result = '';
+      let error = '';
       rufo.on('error', err => {
         console.warn(err);
         vscode.window.showErrorMessage(`couldn't run ${this.exe} '${err.message}'`);
@@ -77,9 +82,15 @@ export default class Rufo {
       rufo.stdout.on('data', data => {
         result += data.toString();
       });
+      rufo.stderr.on('data', data => {
+        console.warn(`Rufo STDERR: ${data}`);
+        error += data.toString();
+      });
       rufo.on('exit', code => {
         if (code) {
-          vscode.window.showErrorMessage(`Rufo failed with exit code: ${code}`);
+          const cleanedError = cleanUpError(error)
+          const msg = cleanedError.length ? cleanedError : `Rufo failed with exit code: ${code}`;
+          vscode.window.showErrorMessage(msg);
           return reject();
         }
         resolve(result);